Split tests by domains
