+++ tests/test_protocol_basics.py
@@ -1,0 +1,68 @@
+# Copyright 2021 Google LLC.
+# Copyright (c) Microsoft Corporation.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from _helpers import *
+
+
+# Tests for "handle an incoming message" error handling, when the message
+# can't be decoded as known command.
+# https://w3c.github.io/webdriver-bidi/#handle-an-incoming-message
+
+@pytest.mark.asyncio
+async def test_binary(websocket):
+    # session.status is used in this test, but any simple command without side
+    # effects would work. It is first sent as text, which should work, and then
+    # sent again as binary, which should get an error response instead.
+    command = {"id": 1, "method": "session.status", "params": {}}
+
+    text_msg = json.dumps(command)
+    await websocket.send(text_msg)
+    resp = await read_JSON_message(websocket)
+    assert resp['id'] == 1
+
+    binary_msg = 'text_msg'.encode('utf-8')
+    await websocket.send(binary_msg)
+    resp = await read_JSON_message(websocket)
+    assert resp == {
+        "error": "invalid argument",
+        "message": "not supported type (binary)"}
+
+
+@pytest.mark.asyncio
+async def test_invalid_json(websocket):
+    message = 'this is not json'
+    await websocket.send(message)
+    resp = await read_JSON_message(websocket)
+    assert resp == {
+        "error": "invalid argument",
+        "message": "Cannot parse data as JSON"}
+
+
+@pytest.mark.asyncio
+async def test_empty_object(websocket):
+    command = {}
+    await send_JSON_command(websocket, command)
+    resp = await read_JSON_message(websocket)
+    assert resp == {
+        "error": "invalid argument",
+        "message": "Expected unsigned integer but got undefined"}
+
+
+@pytest.mark.asyncio
+async def test_session_status(websocket):
+    command = {"id": 5, "method": "session.status", "params": {}}
+    await send_JSON_command(websocket, command)
+    resp = await read_JSON_message(websocket)
+    assert resp == {"id": 5, "result": {"ready": True, "message": "ready"}}
